refactor(desktop): split start-session types and helper utilities
Extract type declarations into start-session.types.ts and utility functions into start-session-support.ts while keeping orchestration flow in start-session.ts.

No functional behavior changes; this is a structural refactor to improve maintainability and file structure.
